stockbt/requirements.txt
torch>=1.9.0
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.4.0
scipy>=1.7.0
alpaca-py>=0.8.0
alpaca-trade-api>=2.3.0
scikit-learn>=1.0.0 